+++ .env
@@ -1,6 +1,6 @@
 
-VUE_APP_REGISTRY_URL=
-VUE_APP_AGGREGATOR_URL=
+VUE_APP_REGISTRY_URL=http://localhost:9090/
+VUE_APP_AGGREGATOR_URL=http://localhost:7070/
 VUE_APP_LOGIN_URL=
 VUE_APP_LOGOUT_URL=
 VUE_APP_JWT_AUDIENCE=
